+++ src/alpaca/common/enums.py
@@ -11,7 +11,6 @@ class BaseURL(str, Enum):
     DATA = "https://data.alpaca.markets"
     DATA_SANDBOX = "https://data.sandbox.alpaca.markets"
     MARKET_DATA_STREAM = "wss://stream.data.alpaca.markets"
-    OPTION_DATA_STREAM = "wss://stream.data.alpaca.markets"  # Deprecated: use MARKET_DATA_STREAM instead!
     TRADING_STREAM_PAPER = "wss://paper-api.alpaca.markets/stream"
     TRADING_STREAM_LIVE = "wss://api.alpaca.markets/stream"